requirements.txt
python-barcode
weasyprint
flask
click
twisted
sqlsoup
sqlalchemy
